Set tanLambda range to -8 to +8

+++ L1Trigger/TrackFindingTMTT/python/TMTrackProducer_Defaults_cfi.py
@@ -460,7 +460,7 @@ TMTrackProducer_params = cms.PSet(
     MiniHT_z0Bits = cms.uint32(10),
     MiniHT_z0Range  = cms.double(60),
     MiniHT_tanlambdaBits = cms.uint32(14),
-    MiniHT_tanlambdaRange = cms.double(32),
+    MiniHT_tanlambdaRange = cms.double(16),
     
     #======= SimpleLR digi parameters ========
     SLR_skipTrackDigi = cms.bool( False ), # Optionally skip track digitisation if done internally inside fitting code.
